refactor(image-loader): remove extended image loader framework and cleanup related utilities
Removed the extended image loader framework including definitions, utilities, and redundant type declarations, as they are no longer required. Centralized loader-specific configurations and streamlined exports to reduce complexity and improve maintainability. Updated import paths and removed dependencies to align with the simplified loader architecture.
