+++ packages/image-loader/src/loaders/extended.ts
